requirements.txt
arviz
matplotlib
numpy
pandas
pyro-ppl
statsmodels
torch
xarray
